FAC-1295: Add answerRange exist check to AnswerRangePersistenceJpaAdapter

+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/adapter/out/persistence/answerrange/AnswerRangePersistenceJpaAdapter.java
@@ -174,6 +174,9 @@ public class AnswerRangePersistenceJpaAdapter implements
 
     @Override
     public void delete(long id, long kitVersionId) {
+        if(!repository.existsByIdAndKitVersionId(id, kitVersionId))
+            throw new ResourceNotFoundException(ANSWER_RANGE_ID_NOT_FOUND);
+
         repository.deleteByIdAndKitVersionId(id, kitVersionId);
     }
 }
